+++ state/s3.go
@@ -1,4 +1,4 @@
-package s3
+package state
 
 import (
 	"context"
