+++ state/s3_test.go
@@ -1,3 +1,3 @@
-package s3
+package state
 
 // TODO: tests for the s3 package
